requirements.txt
exifread
networkx
numpy
utm
PyYAML
six
loky==1.2.1
multiprocessing
opencv-contrib-python

